Updated FileFinder.py
Restructured the functions in a slightly more organized way, and merging helpers into their accompanying main function.

Added playtime, which returns a dict with the structure `{ [subject] : time}`. Time is in seconds.
It finds all the playtime excel files and grabs the "offset" (total playtime) from each. Future analysis programs requires the total time, so this reduces the run-time for those programs.
A consideration is, if it's worthwhile to calculate the time per utterance in some initial setup, for ease of access in future analyses.