Update UttersAnalysis.py
Old one for some reason didn't have dropna() which still can happen in our files occasionally, so that was throwing errors. I also moved punct + transtab to the outside; my local copy has it like that for a long time so I'm not sure why this one didn't, but there's no need to re-create it every time for each new file.